Move supplemental search to correct place
