+++ solution/backend/regcore/search/views.py
@@ -1,10 +1,14 @@
 from datetime import date
-from rest_framework import generics, serializers
+from rest_framework import generics, serializers, viewsets
 from django.db import models
+from django.contrib.postgres.search import SearchHeadline, SearchQuery, SearchVector, SearchRank
 
 from regcore.models import Part
 from .models import SearchIndex
 from django.contrib.postgres.search import SearchQuery, SearchRank, SearchHeadline
+
+from supplemental_content.models import AbstractSupplementalContent
+from supplemental_content.serializers import FlatSupplementalContentSerializer
 
 
 class SearchViewSerializer(serializers.ModelSerializer):
@@ -37,3 +41,40 @@ class SearchView(generics.ListAPIView):
             )\
             .order_by('-rank')\
             .values("type", "content", "headline", "label", "parent", "part__document__title", "part__title", "part__date")
+
+
+class SupplementalContentSearchViewSet(viewsets.ReadOnlyModelViewSet):
+    serializer_class = FlatSupplementalContentSerializer
+
+    def get_queryset(self):
+        query = self.request.query_params.get('q')
+        search_type = 'plain'
+        cover_density = False
+        if (query.startswith('"')) and query.endswith('"'):
+            search_type = 'phrase'
+            cover_density = True
+
+        return AbstractSupplementalContent.objects.filter(approved=True).annotate(rank=SearchRank(
+                SearchVector('supplementalcontent__name', weight='A', config='english')
+                + SearchVector('supplementalcontent__description', weight='A', config='english'),
+                SearchQuery(query, search_type=search_type, config='english'), cover_density=cover_density))\
+            .filter(rank__gte=0.2) \
+            .annotate(
+                nameHeadline=SearchHeadline(
+                    "supplementalcontent__name",
+                    SearchQuery(query, search_type=search_type, config='english'),
+                    start_sel='<span class="search-highlight">',
+                    stop_sel='</span>',
+                    config='english'
+                ),
+                descriptionHeadline=SearchHeadline(
+                    "supplementalcontent__description",
+                    SearchQuery(query, search_type=search_type, config='english'),
+                    start_sel='<span class="search-highlight">',
+                    stop_sel='</span>',
+                    config='english'
+                )
+            )\
+            .order_by('-rank') \
+            .prefetch_related('locations')\
+            .prefetch_related('category').select_subclasses("supplementalcontent")
