+++ solution/backend/regcore/urls.py
@@ -23,7 +23,6 @@ from regcore.v3views import (
     PartSectionsViewSet,
     PartSubpartsViewSet,
     SubpartContentsViewSet,
-    SupplementalContentSearchViewSet
 )
 
 
@@ -76,9 +75,6 @@ urlpatterns = [
         })),
         path("title/<title>/part/<part>/version/<version>/subpart/<subpart>/toc", SubpartContentsViewSet.as_view({
             "get": "retrieve",
-        })),
-        path("supplemental_content/search", SupplementalContentSearchViewSet.as_view({
-            "get": "list",
         })),
     ])),
 ]
