+++ solution/backend/regcore/v3views.py
@@ -1,4 +1,3 @@
-from django.contrib.postgres.search import SearchHeadline, SearchQuery, SearchVector, SearchRank
 from django.shortcuts import get_object_or_404
 
 from rest_framework import viewsets
@@ -17,9 +16,6 @@ from regcore.serializers import (
     PartSubpartsSerializer,
     SubpartContentsSerializer,
 )
-
-from supplemental_content.models import AbstractSupplementalContent
-from supplemental_content.serializers import FlatSupplementalContentSerializer
 
 
 class MultipleFieldLookupMixin(object):
@@ -110,40 +106,3 @@ class SubpartContentsViewSet(PartPropertiesViewSet):
         context = super().get_serializer_context()
         context["subpart"] = self.kwargs.get("subpart")
         return context
-
-class SupplementalContentSearchViewSet(viewsets.ReadOnlyModelViewSet):
-    serializer_class = FlatSupplementalContentSerializer
-
-    def get_queryset(self):
-        query = self.request.query_params.get('q')
-        search_type = 'plain'
-        cover_density = False
-        if (query.startswith('"')) and query.endswith('"'):
-            search_type = 'phrase'
-            cover_density = True
-
-        return AbstractSupplementalContent.objects.filter(approved=True).annotate(rank=SearchRank(
-                SearchVector('supplementalcontent__name', weight='A', config='english')
-                + SearchVector('supplementalcontent__description', weight='A', config='english'),
-                SearchQuery(query, search_type=search_type, config='english'), cover_density=cover_density))\
-            .filter(rank__gte=0.2) \
-            .annotate(
-                nameHeadline=SearchHeadline(
-                    "supplementalcontent__name",
-                    SearchQuery(query, search_type=search_type, config='english'),
-                    start_sel='<span class="search-highlight">',
-                    stop_sel='</span>',
-                    config='english'
-                ),
-                descriptionHeadline=SearchHeadline(
-                    "supplementalcontent__description",
-                    SearchQuery(query, search_type=search_type, config='english'),
-                    start_sel='<span class="search-highlight">',
-                    stop_sel='</span>',
-                    config='english'
-                )
-            )\
-            .order_by('-rank') \
-            .prefetch_related('locations')\
-            .prefetch_related('category').select_subclasses("supplementalcontent")
-
